Updated to show amount_payable and adjusted for string format

+++ app/src/main/java/com/frisky/icebreaker/core/structures/OrderSummary.java
@@ -12,7 +12,7 @@ public class OrderSummary {
     private double totalAmount;
 
     public OrderSummary(String restaurantID, Uri restaurantImage, String restaurantName, String sessionID,
-                        String endTime, int totalAmount) {
+                        String endTime, double totalAmount) {
         this.restaurantImage = restaurantImage;
         this.restaurantID = restaurantID;
         this.restaurantName = restaurantName;
